Package changes, major blueprint adaptions
tl;dr
This commit is way to big and contains major (breaking) changes.

Discovery of microservices now works with a settings blueprint in mib
itself. This has the advantage, that the blueprint can be adapted to
individual needs and at the same time adding microservices automatically
is still possible.
With the new approach, a new developer can do an initial clone and get a
ready-to-work-with infrastructure within minutes without having domain
dependencies in mib initially (because all of these dependencies are
generated upon gradle task execution).

MIB classes still contained domain package naming. Removed that to get
working code right from generation again.